Install eodag from pip
The conda-forge package of eodag brings optional dependencies, among them usgs. We're installing a specific commit of usgs with a patch for the token-based authentication which hasn't been included in a release yet. The version number in this commit has not been updated so pip won't install the package from the commit since the same version is already installed from the conda-forge eodag package. Installing eodag from PyPI avoids this problem.